+++ linen-repository/src/main/scala/x7c1/linen/database/struct/AccountRecord.scala
@@ -3,7 +3,8 @@ package x7c1.linen.database.struct
 import android.database.Cursor
 import x7c1.linen.repository.date.Date
 import x7c1.wheat.macros.database.{TypedCursor, TypedFields}
-import x7c1.wheat.modern.database.{EntityIdentifiable, Insertable, SingleWhere}
+import x7c1.wheat.modern.database.selector.Identifiable
+import x7c1.wheat.modern.database.{Insertable, SingleWhere}
 
 
 trait AccountRecord extends TypedFields {
@@ -22,7 +23,7 @@ object AccountRecord {
   }
 }
 
-trait AccountIdentifiable[A] extends EntityIdentifiable[A, Long]
+trait AccountIdentifiable[A] extends Identifiable[A, Long]
 
 case class AccountParts(
   nickname: String,
