--- wheat-modern/src/main/scala/x7c1/wheat/modern/database/RecordIdentifiable.scala
@@ -1,10 +1,0 @@
-package x7c1.wheat.modern.database
-
-trait RecordIdentifiable[A]{
-  type ID
-  def idOf(target: A): ID
-}
-
-trait EntityIdentifiable[A, X] extends RecordIdentifiable[A]{
-  override type ID = X
-}
