+++ wheat-modern/src/main/scala/x7c1/wheat/modern/database/selector/CanIdentify.scala
@@ -1,0 +1,10 @@
+package x7c1.wheat.modern.database.selector
+
+trait CanIdentify[A]{
+  type ID
+  def idOf(target: A): ID
+}
+
+trait Identifiable[A, X] extends CanIdentify[A]{
+  override type ID = X
+}
